modularize Step1, Step8, and Step9 as callable functions
