+++ Step6A_SHAP_track.py
@@ -8,7 +8,7 @@ import pandas as pd
 import matplotlib
 matplotlib.use('Agg') 
 
-from Step2_trackmodel import TrackNet
+from Step2A_trackmodel import TrackNet
 from Config import *
 
 # === Config ===
